improve encoder performance
While spinning the encoded fast I was able to cause it to lock up repeatedly.  I looked at the example code and noticed they had a way of detecting direction as a 1 or -1 and I decided to use that to increment or decrement the lit up number position. Also made a helper function for turning off all the lights, I think ideally we'd just turn off the last light instead of doing all of them but we can fix that later.